requirements.txt
absl-py==0.15.0
argcomplete==1.12.3
asgiref==3.3.4
astroid==2.4.2
astunparse==1.6.3
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.10.0
bitstring==3.1.9
cached-property==1.5.2
cachetools==4.2.4
caer==2.0.8
certifi==2020.11.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.4
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
clang==5.0
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
cryptography==36.0.0
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.1
django-import-export==2.7.1
djangorestframework==3.13.1
docxcompose==1.3.4
docxtpl==0.15.1
ecdsa==0.17.0
entrypoints==0.3
esptool==3.2
et-xmlfile==1.1.0
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.41.1
gunicorn==20.1.0
h5py==3.1.0
httplib2==0.20.2
idna==2.10
importlib-metadata==4.6.3
ipykernel==6.6.0
ipython==7.30.1
isort==5.6.4
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.7.1
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
mathparse==0.1.2
matplotlib==3.4.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mediapipe==0.8.8.1
MouseInfo==0.1.3
mpmath==1.2.1
mypy==0.910
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.19.5
oauthlib==3.1.1
odfpy==1.4.1
opencv-contrib-python==4.5.4.58
opencv-python==4.5.3.56
openpyxl==3.0.9
opt-einsum==3.3.0
pandas==1.3.4
parso==0.8.3
pickleshare==0.7.5
pid==3.0.4
Pillow==8.3.1
pip-chill==1.0.1
prompt-toolkit==3.0.24
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.10.0
pylint==2.6.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
pyserial==3.5
python-dateutil==2.8.2
python-docx==0.8.11
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.1
pywin32==301
PyYAML==3.13
pyzmq==22.3.0
qrcode==7.3
reedsolo==1.5.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.1
six==1.15.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.24
sqlparse==0.4.1
sympy==1.9
tablib==3.1.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==2.0.1
whitenoise==5.3.0
wikipedia==1.4.0
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==3.0.1
xlwings==0.25.3
xlwt==1.3.0
zipp==3.5.0
